requirements.txt
Flask==3.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
intuit-oauth==1.2.4
cryptography==42.0.0
SQLAlchemy==2.0.25
requests==2.31.0
python-dotenv==1.0.0
PyPDF2==3.0.1
reportlab==4.0.9
boto3==1.34.0
